separate SortHelper
